feat: move help page components to HelpComponents folder
