+++ src/HelpPage/HelpPage.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { Container } from '@material-ui/core';
 import styles from '../common/styles/styles';
-import HelpContent from './HelpContent';
+import HelpContent from './HelpComponents/HelpContent';
 
 export default function HelpPage() {
   const commonStyles = styles();
